WIP(@gve/bot-middleware): manually load Google credentials
affects: @gve/bot-middleware

Google services, e.g., Dialogflow, require that a secrets file
sit on the server with the code. This
requires you to bake the secrets
into the Docker image, which isn't good. Instead, you can provide a
JSON string in an
environment variable and load them manually when creating the client.